Continued to work on color-themes, renamed background and surface color
Background color was renamed from "colorBackgroundFloating" to android:colorBackground to match material guidelines. In order to reference background color, you have to write "?android:attr/colorBackground" instead of "?attr/android:colorBackground". This is a well known bug and android fixes this at runtime, but Android Studio shows it as an error, so lets just write the different style of reference.

colorOnBackground for background_light was renamed to colorSurface, because it matches the material guidelines better than "colorOnBackground", which is reserved for text on background (unused) also added "colorOnSurface" for text on surface-colored-elements

removed "textFillColor" please use text colors apropriate to the background color of the elements you are using for ex: "colorOnSecondary" for text on an element with color "colorSecondary" and so on...

for further information please look into themes.xml Theme.LGG_Frontend (the default theme) which provides clearer information on our color-theme choices